Move VS Code to the root
