+++ src/browser/pages/update.html
@@ -8,14 +8,14 @@
     />
     <meta http-equiv="Content-Security-Policy" content="style-src 'self'; manifest-src 'self'; img-src 'self' data:;" />
     <title>code-server</title>
-    <link rel="icon" href="{{BASE}}/static-{{COMMIT}}/src/browser/media/favicon.ico" type="image/x-icon" />
+    <link rel="icon" href="{{BASE}}/static/{{COMMIT}}/src/browser/media/favicon.ico" type="image/x-icon" />
     <link
       rel="manifest"
-      href="{{BASE}}/static-{{COMMIT}}/src/browser/media/manifest.json"
+      href="{{BASE}}/static/{{COMMIT}}/src/browser/media/manifest.json"
       crossorigin="use-credentials"
     />
-    <link rel="apple-touch-icon" href="{{BASE}}/static-{{COMMIT}}/src/browser/media/code-server.png" />
-    <link href="{{BASE}}/static-{{COMMIT}}/dist/app.css" rel="stylesheet" />
+    <link rel="apple-touch-icon" href="{{BASE}}/static/{{COMMIT}}/src/browser/media/code-server.png" />
+    <link href="{{BASE}}/static/{{COMMIT}}/dist/app.css" rel="stylesheet" />
     <meta id="coder-options" data-settings="{{OPTIONS}}" />
   </head>
   <body>
@@ -33,6 +33,6 @@
         </div>
       </div>
     </div>
-    <script src="{{BASE}}/static-{{COMMIT}}/dist/register.js"></script>
+    <script src="{{BASE}}/static/{{COMMIT}}/dist/register.js"></script>
   </body>
 </html>
